The real hall is not a simple box.  The first set of commits towards a more general shape.
To minimize disruption, keep a formal HallAir volume as G4Box.  (Using
an ExtrudedSolid would require to rotate every existing volume to the
new coordinate system orientation.)  Place ExtMonFNAL and ProtonBeamDump
inside HallAir.  Part of the formal HallAir will be filled with dirt
(so that the remaining air won't be a box any more).  Also wall concrete
still needs to be re-created.