Changes to work with root v6
Requires matching changes in the caller (for example ce.C).

1)
In the original code the variables acan and rcan (both
of type TCanvas*) were local variables within functions.
In root v5 these were visible to the calling routine. In
root v6 this is no longer true.  To make them visible
to the caller, make them public member data.

2)
Change some TH1::Draw calls to DrawCopy calls; or else the histogram
on the TCanvas goes away as soon as the underlying TH1 object goes out
of scope.  So it's not possible to look at the histograms on the
screen.  With DrawCopy the histograms persist on the TCanvas after
the calling script completes.

3)
The clang compiler chokes on:
snprintf(ctext,80,"mcmid.t0%1695>%f",500.);
It interprets the first % as the start of format conversion
operator; it gives a warning and produces incorrect output.
The fix is to double up the first percent sign (the % format
conversion (ie the second %) prints itself and takes no argument).