Changes in preparation for art 3; all compatible with art 2.
The vast majority of changes are removing as many
'RandomNumberGenerator::getEngine' calls as possible.  The one place
where calling 'getEngine' is still necessary is for MixFilter detail
classes; that will be fixed for art 3.01.

Any of the random-number distributions that implicitly called
getEngine now require the random-number engine to be passed in as an
argument.