Fix random-number problem with internal-conversion guns.
There were two instances of random-number distributions that were
function-local static objects.  Although that is not a good design, in
general, worse is that the distributions depend on per-event
calculations.  Static variables are initialized only once, and
therefore only the very first event was relevant in creating the
distribution.

In principle, this means that the calculations for internal
conversions have been incorrect, the extent to which is unknown.  When
RPC/RMC internal conversions are recalculated, one can expect the
results to be a bit different.