Fix StepInstanceName data race.
The construction of static variables is guaranteed to be thread-safe.
However, the operations performed on those variables is not
necessarily thread-safe.  For example:

  static auto const values = step_instance_names(b, e);
  return values;

is thread-safe because the static variable 'values' is initialized
only once, and the C++ language guarantees thread-safe creation of the
variable.  However, the following:

  static std::vector<StepInstanceName> values;
  values = step_instance_names(b, e);  // MT error!
  return values;

is not thread-safe because 'values' is updated outside of the
protected static-variable-initialization region.