Logical Volume Sensitive Detectors (LVSD) have never worked correctly in MT mode.  This commit fixes the immediate problem but the code remains fragile.  Issue #690 in the Offline GitHub issue tracker has been created to remove the fragility.
The immediate problem is that in MT mode there are many SensitiveDetectorHelper
objects and the member function:

   void SensitiveDetectorHelper::instantiateLVSDs(const SimpleConfig& config);

correctly instantiates the objects but stores the pointers in the wrong
SensitiveDetectorHelper object.  They are stored in the instance that lives in
the Mu2eG4MTRunManager object, not in the object that lives in the appropriate
Mu2eG4WorkerRunManager instance.

The solution is to leave the instantion as is but to move the storage of pointers into:

  void SensitiveDetectorHelper::registerSensitiveDetectors();

which is called at the right time using the right SensitiveDetectorHelper instance.

As part of the fix I made it manifest that instantiateLVSDs does not register
any pointers; I also  moved an invariant compution out of a loop.