Changed the signature of the c'tor of PhysicalVolumeMultiHelper to take it's argument by const* not const reference.  Plus the follow-on changes in each of the callers.
Each of the callers held a pointer to the object and the previous interface forced
a needless transformation from pointer to reference and back to pointer.
The pointer is allowed to have a value of nullptr, which caused sanitize
error messages when any of the callers derefenced the pointer to pass
it by reference.