This commit addresses the deprecation of art::Event::get(art::ProductID) from the interface of art::Event .
The return type of the function Mu2eG4Inputs::inputSimParticles can no longer be an art::Handle<SimParticleCollection> because there is no way that I know of to construct such an handle in all cases with the tools available in art v3.11  .   To address this, I defined a small helper class to serve as the new return type: Mu2eG4Inputs::InputSimsInfo .

The two classes that use this function need two things that they can get from the Handle: the data product itself, and the art::ProductID of the data product.  The two data members of the new class are a smart pointer to the data product itself and the art::ProductID.  The smart pointer uses the class template maybe_ref which comes from the cetlib library; it behaves like a minimal handle class template but returns it??s payload by reference not pointer.

I considered std::optional instead of maybe_ref.  I chose maybe_ref because it throws an exception if one forgets to check isValid and tries to use the referent from an invalid object; on the other, when I tested std::optional it seg faulted when tried to use the referent from an invalid object.