Update GeometryService to use validated fhicl.
In the process I changed one more thing: it no longer caches the full parameter set.
The only part of the cached parameter set that is ever used is the simulatedDetector
component; so I cached only that and renamed a few things to match the change in
functionality.  And changed the code that uses it to use the new name.