In GeometryService, support the model in which the bfield description file has been separated from the geometry file.  Remove support for the model in which both are all in one file.
Update default configuration of GeometryService to include the definition
of the default bfield file; this is in the Services.Core section of fcl/standardServices.fcl.

Remove bfield from the current, named geometry file:
 - Mu2eG4/geom/geom_2021_PhaseI_v02.txt

Two of the bfgeom files need to have one extra line added.  The others
include one of these and pick up the extra line transitively.

I will address follow on issues in a separate commit.