Quick fix for the crashes in g4test_03MT.fcl and transportOnly.fcl introduced by Offline PR #1183.
The fix is to remove all hit processing downstream of g4run and then
edit ReadBack_module.cc to skip steps when it's inputs are missing.

The full fix will be to remove unneeded parts of ReadBack_module.cc
in their entirety, prune configuration that is no longer needed,
switch to validated fcl and update about 15 .fcl files (some of which can
be deleted).