Change the pdt_ data member type:
-    mutable GlobalConstantsHandle<ParticleDataList> pdt_;
+    ParticleDataList const& pdt_;

When the data member is a handle it (very indirectly) causes the dictionary
maker for RecoDataProducts to fail.  If you follow the include chain it eventually
includes fhiclcpp/ParameterSet.h which has content that the dictionary maker does
not know how to deal with.  It is not necessary for the data member to be a handle
and this change makes it a const& to ParticleDataList.  All other changes trivially
follow from this.