ExtMonFNAL: a small geometry change
This commit removes the obsolete _horizontalLength member of the
collimator class.  This led to a re-definition of filterExitInMu2e(),
which is a reference point for placing the detector.  All in all, a
comparison of gdml files before and after the change shows that ExtMon detector
shifted by -142.177529457077 mm in Z, and by cm-level in X and Y:

-        <position name="ExtMonStackMotherUp_pos" unit="mm" x="641.604080528331" y="10.7149794556072" z="1219.4789351695"/>
+        <position name="ExtMonStackMotherUp_pos" unit="mm" x="645.899038646913" y="25.9382927782449" z="1077.30140571242"/>

same shifts for other related volumes.