WIP: activate the new code for ExtMon exit collimator placement, geometry changed.
Collimator2 is aligned to the filter magnet, so it shifted in the same
way as the magnet in the previous commit:

-        <position name="collimator2Mother_pos" unit="mm" x="-1070.88591036244" y="-2761.49535163724" z="-31264.4948907125"/>
+        <position name="collimator2Mother_pos" unit="mm" x="-1071.35546568122" y="-2764.1503543482" z="-31264.3802518035"/>

There are more similar shifts in other objects referenced to the
filter exit - the spectrometer magnet, stack mothers, and virtual
detectors.  There are two confusing changes in mu2e.gdml

-      <position name="ExtMonFNALColl2Shielding_pos" unit="mm" x="-25311.4948907125" y="-157.385910362437" z="-145.295351637237"/>
+      <position name="ExtMonFNALColl2Shielding_pos" unit="mm" x="-25311.3802518035" y="-157.855465681215" z="-147.950354348197"/>

-      <position name="ExtMonDetectorRoom_pos" unit="mm" x="597.588837334626" y="-145.295351637236" z="2676.53121902264"/>
+      <position name="ExtMonDetectorRoom_pos" unit="mm" x="597.105490381774" y="-147.950354348196" z="2676.53121939323"/>

but those are shift NOT of the shielding wall and the room, but of their constituent volumes
used to construct the subtraction solids.  The room and the wall themselves did not move.